+++ src/ReverseProxy/Configuration/DependencyInjection/BuilderExtensions/IReverseProxyBuilderExtensions.cs
@@ -50,7 +50,14 @@ namespace Microsoft.ReverseProxy.Configuration.DependencyInjection
         public static IReverseProxyBuilder AddConfigManager(this IReverseProxyBuilder builder)
         {
             builder.Services.TryAddSingleton<IProxyConfigManager, ProxyConfigManager>();
-            builder.Services.TryAddSingleton<IProxyAppState, ProxyAppState>();
+
+            // Avoid registering several IProxyAppState implementations.
+            if (!builder.Services.Any(d => d.ServiceType == typeof(IProxyAppState)))
+            {
+                builder.Services.AddSingleton<ProxyAppState>();
+                builder.Services.AddSingleton<IProxyAppState>(p => p.GetRequiredService<ProxyAppState>());
+                builder.Services.AddSingleton<IProxyAppStateSetter>(p => p.GetRequiredService<ProxyAppState>());
+            }
             return builder;
         }
